Pass layer object instead of layer name to load() to avoid bug
This change avoids a bug that we fixed in ArcGIS Pro 2.6.  Users running this script in 2.5 will run into it and have problems with chunking.

In this script, we use MakeFeatureLayer with a where clause to select ObjectID ranges in order to chunk up the data into small enough parts that don’t exceed the desired limits.  The bug was that when we called the load() method on the NA solver objects, load() was ignoring the where clause on the feature layer, but only if the feature layer is referenced using its string name.  If you reference it as a layer object, it works fine, so that’s the fix for the script.