Fix logic for selecting destinations based on chunk OIDs and cutoffs
This logic was all messed up, and there were codepaths that led to nothing being selected, and that messed up load() later.

To fix it, first select the correct destinations for the chunk based on OIDs.  Then, if the user has designated a cutoff, perform a further selection to select only the OIDs within the cutoff.  If at this point nothing is selected, then there are no valid destinations for this chunk. Before calling load() make sure there is a valid destination layer.  If there isn't, just end the chunk and move on.