Future proofing for Pro 3.0 release
Suppress tool warnings when setting searchTolerance and searchToleranceUnits if the network data source is an older portal that does not support these settings.  These settings are always set using the value in the config file, but older portals do not support this setting and always use their default configuration.  This will generate a warning every time the tool is run, which we don't want.  Just suppress the warning.

Note that in 2.9, the searchTolerance and searchToleranceUnits could be successfully set, but the service would simply ignore the values without informing the user. This has been fixed in 3.0, but it causes minor fallout in tools like this.